libretro: fix compile on android (#68)

+++ io/iocore.h
@@ -3,8 +3,8 @@
 
 typedef void (*FNIORESET)(const NP2CFG *pConfig);
 typedef void (*FNIOBIND)(void);
-typedef	void (IOOUTCALL *IOOUT)(UINT port, UINT val);
-typedef	UINT (IOINPCALL *IOINP)(UINT port);
+typedef void (IOOUTCALL *IOOUT)(UINT port, UINT8 val);
+typedef UINT8 (IOINPCALL *IOINP)(UINT port);
 
 #include	"lsidef.h"
